backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
pydantic
python-dotenv
passlib[bcrypt]
python-jose
aiobotocore
websockets
PyJWT
